+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 11.6.0b4 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 11.6.0b3 (2024-04-09)
 
 ### Features Added